requirements.txt
boto3
ophis-py